Rename : 술도감 화면 네이밍 변경 및 OrangeColorButton 패키지 이동
1. SpiritsCollection -> LiquorCollection
2. 변경 범위 : Fragment, ViewModel, xml, navigation
3. OrangeColorButton 패키지 수정 ui.component -> ui.component.button
