+++ app/src/main/java/com/goody/dalda/ui/component/SimpleMessageDialog.kt
@@ -16,6 +16,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
+import com.goody.dalda.ui.component.button.OrangeColorButton
 import com.goody.dalda.ui.theme.DaldaTextStyle
 
 @Composable
