+++ app/src/main/java/com/goody/dalda/ui/home/component/LoginBanner.kt
@@ -3,14 +3,13 @@ package com.goody.dalda.ui.home.component
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.goody.dalda.ui.component.AutoResizedText
-import com.goody.dalda.ui.component.OrangeColorButton
+import com.goody.dalda.ui.component.button.OrangeColorButton
 import com.goody.dalda.ui.theme.DaldaTextStyle
 
 @Composable
