+++ app/src/main/res/layout/fragment_liquor_collection.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.spirits_collection.SpiritsCollectionFragment">
+    tools:context=".ui.spirits_collection.LiquorCollectionFragment">
 
     <TextView
         android:id="@+id/text_spirits_collection"
