+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">Dalda</string>
     <string name="title_home">홈</string>
-    <string name="title_spirits_collection">술 도감</string>
+    <string name="title_liquor_collection">술 도감</string>
     <string name="title_undefined">미정</string>
     <string name="icon_menu">메뉴</string>
     <string name="text_welcome_comment">%1$s 님,\n안녕하세요!</string>
